+++ frontend/src/pages/AccountSelection.tsx
@@ -4,6 +4,7 @@ import { AccountContextProps, Account, Accounts } from '../types';
 import { AccountContext } from '../context/AccountContext';
 import Button from '../components/ui/Button';
 import { useNavigate } from 'react-router-dom';
+import TwitterLogin from './TwitterLogin';
 
 const AccountSelection = () => {
 	const [accounts, setAccounts] = useState<Accounts>([]);
@@ -54,6 +55,14 @@ const AccountSelection = () => {
 		);
 	});
 
+	const handleAddAccount = (e: React.MouseEvent<HTMLButtonElement>) => {
+		e.preventDefault();
+		const windowFeatures = 'left=100,top=100,width=320,height=320';
+		const handle = window.open('https://www.mozilla.org/', '', windowFeatures);
+
+		return <TwitterLogin />;
+	};
+
 	return (
 		<div className="container mx-auto flex justify-center">
 			<div className="mt-8 w-full xl:w-2/5">
@@ -64,7 +73,8 @@ const AccountSelection = () => {
 					<p className="text-center mt-6">Click on "Add Account" to link your first Twitter account.</p>
 				)}
 				<div className="mt-6 flex justify-center">
-					<Button text={'+ Add Account'} />
+					{/*<Button text={'+ Add Account'} />*/}
+					<button onClick={(e) => handleAddAccount(e)}>Add Account</button>
 				</div>
 			</div>
 		</div>
